+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Measurement Table:
 - The neuron name is now quoted in the CSV file export. This makes it more
   robust when commas are used in the name representation (e.g. if annotations).
 
+- XLSX spreadsheet export is now supported through an "Export XLSX" button.
+
 
 ### Bug fixes